API project with swagger integration added
To add easy access to the solution data, I  have added a minimal API with swagger front to the project. This caused some issues to surface, which I have solved. These issues are circular references in the virtual navigational properties when serializing to json and I had forgotten to add class and race type to the character class